service/ec2: aws_default_route_table route compatibility with v0.12
By setting ConfigMode to schema.SchemaConfigModeAttr we ensure that
Terraform v0.12 will permit explicitly assigning an empty list to this
attribute using attribute syntax, vs. requiring this to be given as a
nested block.

That's important for this attribute because not setting the attribute at
all is treated differently than setting it to an empty list.